// backend/src/libs/whatsAppOficial/IWhatsAppOficial.interfaces.ts
export interface IMetaMessageTemplate {
    name: string;
    language: IMetaMessageTemplateLanguage;
    components?: Array<IMetaMessageTemplateComponents> | IMetaMessageTemplateComponents;
}

export interface IMetaMessageTemplateLanguage {
    code: string;
}

export interface IMetaMessageTemplateComponents {
    type: 'header' | 'body' | 'footer' | 'button';
    sub_type?: 'quick_reply' | 'url'
    index?: string;
    parameters: IMetaMessageTemplateComponentsParameters | Array<IMetaMessageTemplateComponentsParameters>;
}

export interface IMetaMessageTemplateComponentsParameters {
    type: 'location' | 'currency' | 'date_time' | 'text' | 'payload' | 'location' | 'url' | 'image' | 'video' | 'document';
    text?: string;
    url?: string;    
    location?: IMetaMessageTemplateComponentsParametersLocation
    currency?: IMetaMessageTemplateComponentsParametersCurrency;
    image?: IMetaMessageTemplateComponentsParametersImage;
    video?: { link: string };
    document?: { link: string; filename?: string };
    date_time?: IMetaMessageTemplateComponentsParametersDateTime;
    payload?: string;
}

export interface IMetaMessageTemplateComponentsParametersLocation {
    latitude: string;
    longitude: string;
    name?: string;
    address?: string;
}

export interface IMetaMessageTemplateComponentsParametersCurrency {
    fallback_value: string;
    code: string;
    amount_1000: number;
}

export interface IMetaMessageTemplateComponentsParametersImage {
    link: string;
}

export interface IMetaMessageTemplateComponentsParametersDateTime {
    fallback_value: string;
}